feature: secure access to data

+++ analyzer/views.py
@@ -5,7 +5,9 @@ from django.contrib import messages
 from django.core.paginator import Paginator, EmptyPage
 
 def index(request, page=1):
-    sequences = Sequence.objects.order_by('-id')
+    if not request.user.is_active:
+        return redirect('user_login')
+    sequences = Sequence.objects.filter(user=request.user.id)
     paginator = Paginator(sequences, 20, 5)
     try:
         sequences = paginator.page(page)
@@ -15,27 +17,33 @@ def index(request, page=1):
 
 
 def create(request):
+    if not request.user.is_active:
+        return redirect('user_login')
     if request.method == 'POST':
         form = SequenceForm(request.POST, request.FILES)
         if form.is_valid():
-            return redirect('read', form.save(request.POST['user']).id)
+            return redirect('read', form.save(request.user.id).id)
     form = SequenceForm()
     return render(request, 'analyzer/sequence_form.html', locals())
 
 
 def read(request, id):
-    sequence = get_object_or_404(Sequence, id=id)
+    if not request.user.is_active:
+        return redirect('user_login')
+    sequence = get_object_or_404(Sequence, id=id, user=request.user.id)
     if sequence.nb_bases is not None:
         sequence.nb_bases = int(sequence.nb_bases)
     return render(request, 'analyzer/read.html', locals())
 
 
 def update(request, id):
-    sequence = get_object_or_404(Sequence, id=id)
+    if not request.user.is_active:
+        return redirect('user_login')
+    sequence = get_object_or_404(Sequence, id=id, user=request.user.id)
     if request.method == 'POST':
         form = SequenceForm(request.POST, request.FILES, instance=sequence)
         if form.is_valid():
-            form.save()
+            form.save(request.user.id)
             return redirect('read', id)
     else:
         form = SequenceForm(instance=sequence)
@@ -43,14 +51,18 @@ def update(request, id):
 
 
 def delete(request, id):
-    sequence = get_object_or_404(Sequence, id=id)
+    if not request.user.is_active:
+        return redirect('user_login')
+    sequence = get_object_or_404(Sequence, id=id, user=request.user.id)
     sequence.delete()
     messages.add_message(request, messages.SUCCESS, 'La séquence ADN a bien été supprimée !')
     return redirect('index')
 
 
 def analyze(request, id):
-    sequence = get_object_or_404(Sequence, id=id)
+    if not request.user.is_active:
+        return redirect('user_login')
+    sequence = get_object_or_404(Sequence, id=id, user=request.user.id)
     sequence.analyse()
     sequence.ratio_g_c_graph()
     sequence.dna_walk_graph()
